article.py: Update to use new models extractor.py: number_of_hashed_references(), percent_of_references_with_a_hash(): Move in from article.py
